+++ api/lib/reducers/akashV0-reducers.js
@@ -18,18 +18,15 @@ function setTransactionSuccess(transaction) {
   return transaction.code ? false : true
 }
 
-function delegationReducer(delegation, validator, network) {
-  const coinLookup = network.getCoinLookup(network, delegation.balance.denom)
-  const delegationCoin = terraV3Reducers.coinReducer(
-    delegation.balance,
-    coinLookup
-  )
+function delegationReducer(delegation, validator) {
   return {
     id: delegation.validator_address,
     validatorAddress: delegation.validator_address,
     delegatorAddress: delegation.delegator_address,
     validator,
-    amount: delegationCoin.amount
+    amount: delegation.balance
+      ? terraV3Reducers.atoms(delegation.balance.amount)
+      : 0
   }
 }