+++ api/lib/reducers/cosmosV0-reducers.js
@@ -166,6 +166,15 @@ function governanceParameterReducer(depositParameters, tallyingParamers) {
   }
 }
 
+function topVoterReducer(topVoter) {
+  return {
+    name: topVoter.name,
+    address: topVoter.operatorAddress,
+    votingPower: topVoter.votingPower,
+    validator: topVoter
+  }
+}
+
 function getValidatorStatus(validator) {
   if (validator.status === 2) {
     return {
@@ -339,6 +348,7 @@ async function balanceV2Reducer(
   lunieCoin,
   stakingDenom,
   delegations,
+  undelegations,
   fiatValueAPI,
   fiatCurrency,
   address
@@ -348,8 +358,12 @@ async function balanceV2Reducer(
     (sum, { amount }) => BigNumber(sum).plus(amount),
     0
   )
+  const undelegatingStake = undelegations.reduce(
+    (sum, { amount }) => BigNumber(sum).plus(amount),
+    0
+  )
   const total = isStakingDenom
-    ? lunieCoin.amount.plus(delegatedStake)
+    ? lunieCoin.amount.plus(delegatedStake).plus(undelegatingStake)
     : lunieCoin.amount
   const fiatValue = await fiatValueAPI.calculateFiatValues(
     [
@@ -392,7 +406,7 @@ function delegationReducer(delegation, validator, active) {
 
 function undelegationReducer(undelegation, validator) {
   return {
-    id: validator.operatorAddress,
+    id: `${validator.operatorAddress}_${undelegation.creation_height}`,
     delegatorAddress: undelegation.delegator_address,
     validator,
     amount: atoms(undelegation.balance),
@@ -497,6 +511,7 @@ function extractInvolvedAddresses(transaction) {
 module.exports = {
   proposalReducer,
   governanceParameterReducer,
+  topVoterReducer,
   tallyReducer,
   validatorReducer,
   blockReducer,