+++ api/lib/schema.js
@@ -10,6 +10,13 @@ const typeDefs = gql`
     CURRENCY
   }
 
+  enum proposalTypeEnum {
+    TEXT
+    COUNCIL
+    TREASURY
+    PARAMETER_CHANGE
+  }
+
   type Tally {
     yes: String # BigNumber
     no: String # BigNumber
@@ -57,9 +64,9 @@ const typeDefs = gql`
   }
 
   type Proposal {
-    id: Int
+    id: String
     networkId: String!
-    type: String
+    type: proposalTypeEnum
     title: String
     description: String
     status: String
@@ -70,6 +77,7 @@ const typeDefs = gql`
     deposit: String # BigNumber
     proposer: String
     validator: Validator
+    beneficiary: String
   }
 
   type Validator {
@@ -126,6 +134,7 @@ const typeDefs = gql`
     chainDenom: String!
     viewDenom: String!
     chainToViewConversionFactor: Float!
+    icon: String
   }
 
   enum CapabilityEnum {
@@ -309,11 +318,32 @@ const typeDefs = gql`
     blockExplorerLink: String
   }
 
+  type TopVoter {
+    name: String!
+    address: String!
+    votingPower: String!
+    validator: Validator
+  }
+
   type GovernanceParameters {
     depositDenom: String
     votingThreshold: Float
     vetoThreshold: Float
     depositThreshold: String # BigNumber
+  }
+
+  type GovernanceLink {
+    title: String
+    link: String
+    type: String
+  }
+
+  type GovernanceOverview @cacheControl(maxAge: 21600) {
+    totalStakedAssets: Float
+    totalVoters: Int
+    treasurySize: Float
+    topVoters: [TopVoter]
+    links: [GovernanceLink]
   }
 
   type Vote {
@@ -340,6 +370,11 @@ const typeDefs = gql`
     properties: String # JSON encoded as it is general purpose and GraphQL doesn't allow for "Object"
   }
 
+  type Session {
+    sessionToken: String!
+    validUntil: String!
+  }
+
   type Subscription {
     blockAdded(networkId: String!): BlockV2
     notificationAdded(addressObjects: [NotificationInput]!): Notification
@@ -349,10 +384,11 @@ const typeDefs = gql`
   }
 
   type Mutation {
-    registerUser(idToken: String!): Boolean
+    registerUser(idToken: String!): Session
     notifications(
       addressObjects: [NotificationInput]!
       notificationType: String
+      pushToken: String
     ): Boolean
   }
 
@@ -409,7 +445,7 @@ const typeDefs = gql`
 
   type Query {
     blockV2(networkId: String!, height: Int): BlockV2
-    proposal(networkId: String!, id: Int!): Proposal
+    proposal(networkId: String!, id: String!): Proposal
     proposals(networkId: String!): [Proposal]
     validators(
       networkId: String!
@@ -418,8 +454,9 @@ const typeDefs = gql`
       popularSort: Boolean
     ): [Validator]
     allDelegators(networkId: String!): [String]
-    vote(networkId: String!, proposalId: Int!, address: String!): Vote
+    vote(networkId: String!, proposalId: String!, address: String!): Vote
     governanceParameters(networkId: String!): GovernanceParameters
+    governanceOverview(networkId: String!): GovernanceOverview
     validator(networkId: String!, operatorAddress: String!): Validator
     networks(experimental: Boolean): [Network]
     network(id: String): Network