+++ app/src/components/account/Paywall.vue
@@ -60,11 +60,10 @@
             <li class="table-cell check">Github Alerts</li>
           </ul>
         </div>
-        <div class="table-span" @click="handleIntercom()">
-          <span>
-            Have any ideas? Share them with us
-            <i class="material-icons notranslate">launch</i>
-          </span>
+        <div class="table-span">
+          <a @click="handleIntercom()">
+            Have any ideas? Share them with us!
+          </a>
         </div>
       </div>
     </div>
@@ -156,12 +155,15 @@ export default {
     },
   },
   watch: {
-    userSignedIn: function () {
-      if (this.userSignedIn) {
-        this.$router.push({
-          name: `notifications`,
-        })
-      }
+    userSignedIn: {
+      immediate: true,
+      handler(userSignedIn) {
+        if (userSignedIn) {
+          this.$router.push({
+            name: `notifications`,
+          })
+        }
+      },
     },
   },
   methods: {