+++ app/src/components/common/DisconnectedBar.vue
@@ -22,7 +22,6 @@ export default {
   },
   methods: {
     reconnect() {
-      console.log("reconnect")
       this.$store.dispatch("reconnect")
     },
   },