+++ app/src/gql/index.js
@@ -4,7 +4,6 @@ import gql from "graphql-tag"
 import store from "../vuex/store"
 
 function getCurrentNetwork() {
-  // console.log(store())
   return store().state.connection.network
 }
 
@@ -40,7 +39,6 @@ export const ValidatorFragment = `
 
 export const AllValidators = () => {
   const currentNetwork = getCurrentNetwork()
-  // console.log(`currentNetwork`, currentNetwork)
   return gql`
     query AllValidators {
       validators(networkId: "${currentNetwork}") {
@@ -137,6 +135,7 @@ export const NetworksAll = gql`
         chainDenom
         viewDenom
         chainToViewConversionFactor
+        icon
       }
       rpc_url
       HDPaths
@@ -149,7 +148,7 @@ export const NetworksAll = gql`
 
 export const NetworksResult = (data) => data.networks
 
-const ProposalFragment = `
+export const ProposalFragment = `
   id
   type
   title
@@ -174,7 +173,7 @@ const ProposalFragment = `
 `
 
 export const ProposalItem = (schema) => gql`
-  query proposal($id: Int!) {
+  query proposal($id: String!) {
     proposal(networkId: "${schema}", id: $id) {
       ${ProposalFragment}
     }
@@ -193,7 +192,7 @@ query governanceParameters {
 `
 
 export const Vote = (schema) => gql`
-query vote($proposalId: Int!, $address: String!) {
+query vote($proposalId: String!, $address: String!) {
   vote(networkId: "${schema}", proposalId: $proposalId, address: $address) {
     option
   }