+++ app/src/initializeApp.js
@@ -1,5 +1,8 @@
 /* istanbul ignore file: really just integrations */
 
+import { Plugins } from "@capacitor/core"
+import * as Sentry from "@sentry/browser"
+const { App: CapacitorApp } = Plugins
 import { listenToExtensionMessages } from "scripts/extension-utils"
 import { checkForNewLunieVersions } from "scripts/check-for-new-lunie-versions"
 import {
@@ -9,9 +12,8 @@ import {
 import config from "src/../config"
 import Router, { routeGuard } from "./router"
 import Store from "./vuex/store"
-import { createApolloProvider } from "src/gql/apollo.js"
-import { Plugins } from "@capacitor/core"
-const { App: CapacitorApp } = Plugins
+import { createApolloProvider, apolloErrorHandler } from "src/gql/apollo.js"
+import { registerForPushNotifications } from "./scripts/pushNotifications"
 import { handleDeeplink, getLaunchUrl } from "./vuex/modules/account"
 
 if (navigator && navigator.serviceWorker) {
@@ -49,11 +51,24 @@ export default async function init(urlParams, env = process.env) {
   const apolloClient = apolloProvider.clients.defaultClient
 
   const store = Store({ apollo: apolloClient })
+
+  // we need to use this custom error handler as we want to use the store in there
+  // we can't pass the store as it would create a circular dependency
+  apolloErrorHandler.onError((error) => {
+    if (error.extensions && error.extensions.code === "UNAUTHENTICATED") {
+      store.dispatch("signOutUser")
+      return
+    }
+    Sentry.captureException(error)
+    console.error(error)
+  })
+
   // we need to set url params before querying for networks because of experimental flag
   setOptions(urlParams, store)
 
   // check if user is signed in
-  await store.dispatch(`listenToAuthChanges`)
+  store.dispatch(`listenToAuthChanges`) // handles Google OAuth changes we are not influencing
+  await store.dispatch(`checkSession`)
 
   // we load the networks first as we need them in the router
   await store.dispatch(`preloadNetworkCapabilities`)
@@ -66,10 +81,17 @@ export default async function init(urlParams, env = process.env) {
     setGoogleAnalyticsPage(to.path)
   })
 
-  CapacitorApp.addListener("appUrlOpen", function (data) {
-    handleDeeplink(data.url, router)
-  })
-  getLaunchUrl(router)
+  registerForPushNotifications(store, router)
+
+  if (config.mobileApp) {
+    CapacitorApp.addListener("appUrlOpen", function (data) {
+      handleDeeplink(data.url, router)
+    })
+    // handling deeplinks when app is opening
+    getLaunchUrl(router)
+  } else {
+    handleDeeplink(window.location, router)
+  }
 
   store.dispatch(`loadLocalPreferences`)
   await store.dispatch(`checkForPersistedNetwork`) // wait until signin