+++ app/tests/unit/specs/components/ActionModal/components/DelegationModal.spec.js
@@ -250,7 +250,7 @@ describe(`DelegationModal`, () => {
       }
       DelegationModal.methods.onSuccess.call(self)
       expect(self.$store.dispatch).toHaveBeenCalledWith(
-        "updateEmailRegistrations"
+        "updateNotificationRegistrations"
       )
     })
   })